Temporarily pin kali box to earlier version
There is currently a bug with the kali v2022.3.1 and v2022.3.2 vagrant
boxes that prevent them from running normally. This commit is meant to
be a temporary solution to use the earlier versions. It is especially
important given that the first club meeting which includes VM setup is
tonight. This commit can be reverted once the issue with the base boxes
is fixed. See https://gitlab.com/kalilinux/build-scripts/kali-vagrant/-/issues/6
for more details about the issue.